修正編輯工具列在某些狀況下會消失的bug
1.修正編輯工具列在某些狀況下會消失的bug
2.修正檔案上傳欄位，避免超過6萬5千則留言後，檔案無法上傳的情形
3.調整討論區代表圖片，使之可自動放。

+++ sql/mysql.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `tad_discuss` (
   `DiscussID` mediumint(9) unsigned NOT NULL auto_increment COMMENT '編號',
   `ReDiscussID` mediumint(9) unsigned NOT NULL default 0 COMMENT '回覆編號',
-  `uid` smallint(6) unsigned NOT NULL default 0 COMMENT '發布者',
+  `uid` mediumint(8) unsigned NOT NULL default 0 COMMENT '發布者',
   `publisher` varchar(255) NOT NULL default '' COMMENT '發布者姓名',
   `DiscussTitle` varchar(255) NOT NULL default '' COMMENT '標題',
   `DiscussContent` text NOT NULL COMMENT '內容',
@@ -30,7 +30,7 @@ PRIMARY KEY (`BoardID`)
 CREATE TABLE `tad_discuss_files_center` (
   `files_sn` smallint(5) unsigned NOT NULL AUTO_INCREMENT COMMENT '檔案流水號',
   `col_name` varchar(255) NOT NULL default '' COMMENT '欄位名稱',
-  `col_sn` smallint(5) unsigned NOT NULL default 0 COMMENT '欄位編號',
+  `col_sn` mediumint(59) unsigned NOT NULL default 0 COMMENT '欄位編號',
   `sort` smallint(5) unsigned NOT NULL default 0 COMMENT '排序',
   `kind` enum('img','file') NOT NULL default 'img' COMMENT '檔案種類',
   `file_name` varchar(255) NOT NULL default '' COMMENT '檔案名稱',
